backend/requirements.txt
psycopg2-binary
openai
tqdm
python-dotenv
